requirements.txt
streamlit==1.40.0
pandas==2.2.3
numpy==2.1.3
plotly==5.24.1
pymongo[srv]==4.10.1
scikit-learn==1.5.2
python-dotenv==1.0.1
certifi
